feat: refactor to address latest review
