+++ package.json
@@ -77,8 +77,8 @@
 		"@zero-tech/zapp-utils": "^0.5.1",
 		"@zero-tech/zero-contracts": "^0.0.7",
 		"@zero-tech/zns-sdk": "0.9.1",
-		"@zero-tech/zui": "0.11.5",
 		"@zero-tech/ztoken-sdk": "^0.0.3",
+		"@zero-tech/zui": "0.11.5",
 		"classnames": "^2.3.1",
 		"formik": "^2.2.9",
 		"moment": "^2.29.4",
