+++ src/features/create-token/Steps/DetailsForm.test.tsx
@@ -4,17 +4,17 @@ import userEvent from '@testing-library/user-event';
 import { fireEvent, render, waitFor, screen } from '@testing-library/react';
 
 import { ZUIProvider } from '@zero-tech/zui/ZUIProvider';
-import { TokenomicsForm, TokenomicsFormProps } from './';
+import { DetailsForm, DetailsFormProps } from './';
 import { CreateTokenFormContext } from '../';
 
 let onSubmit = jest.fn();
 
-const DEFAULT_PROPS: TokenomicsFormProps = {
+const DEFAULT_PROPS: DetailsFormProps = {
 	onClose: jest.fn(),
 };
 
 const DEFAULT_PROVIDER_VALUES = {
-	stepId: 'tokenomics',
+	stepId: 'details',
 	title: 'Create Token',
 	details: {
 		mediaType: undefined,
@@ -30,24 +30,24 @@ const DEFAULT_PROVIDER_VALUES = {
 	onStepUpdate: jest.fn(),
 	onTitleUpdate: jest.fn(),
 	onDetailsChange: jest.fn(),
-	onDetailsSubmit: jest.fn(),
-	onTokenomicsSubmit: onSubmit,
+	onDetailsSubmit: onSubmit,
+	onTokenomicsSubmit: jest.fn(),
 	onLaunchSubmit: jest.fn(),
 };
 
-describe('<TokenomicsForm />', () => {
+describe('<DetailsForm />', () => {
 	beforeEach(() => jest.resetAllMocks());
 
-	test('should correctly validate required token count field', async () => {
-		render (
+	test('should correctly validate required name field', async () => {
+		render(
 			<ZUIProvider>
 				<CreateTokenFormContext.Provider value={DEFAULT_PROVIDER_VALUES}>
-					<TokenomicsForm {...DEFAULT_PROPS} />
+					<DetailsForm {...DEFAULT_PROPS} />
 				</CreateTokenFormContext.Provider>
-			</ZUIProvider>
+			</ZUIProvider>,
 		);
 
-		fireEvent.blur(screen.getByPlaceholderText(/Enter total supply.../i));
+		fireEvent.blur(screen.getByPlaceholderText(/Enter name.../i));
 		fireEvent.click(
 			screen.getByRole('button', {
 				name: 'Next',
@@ -55,20 +55,22 @@ describe('<TokenomicsForm />', () => {
 		);
 
 		await waitFor(() =>
-			expect(screen.getByText('The token count field is required.')).not.toBe(null),
+			expect(
+				screen.getByText('The name field is required.'),
+			).toBeInTheDocument(),
 		);
 	});
 
-	test('should correctly validate required initial wallet address field', async () => {
-		render (
+	test('should correctly validate required symbol field', async () => {
+		render(
 			<ZUIProvider>
 				<CreateTokenFormContext.Provider value={DEFAULT_PROVIDER_VALUES}>
-					<TokenomicsForm {...DEFAULT_PROPS} />
+					<DetailsForm {...DEFAULT_PROPS} />
 				</CreateTokenFormContext.Provider>
-			</ZUIProvider>
+			</ZUIProvider>,
 		);
 
-		fireEvent.blur(screen.getByPlaceholderText(/Enter initial wallet address.../i));
+		fireEvent.blur(screen.getByPlaceholderText(/Enter symbol.../i));
 		fireEvent.click(
 			screen.getByRole('button', {
 				name: 'Next',
@@ -76,28 +78,9 @@ describe('<TokenomicsForm />', () => {
 		);
 
 		await waitFor(() =>
-			expect(screen.getByText('The initial wallet address field is required.')).not.toBe(null),
-		);
-	});
-
-	test('should correctly validate required admin wallet address field', async () => {
-		render (
-			<ZUIProvider>
-				<CreateTokenFormContext.Provider value={DEFAULT_PROVIDER_VALUES}>
-					<TokenomicsForm {...DEFAULT_PROPS} />
-				</CreateTokenFormContext.Provider>
-			</ZUIProvider>
-		);
-
-		fireEvent.blur(screen.getByPlaceholderText(/Enter admin wallet address.../i));
-		fireEvent.click(
-			screen.getByRole('button', {
-				name: 'Next',
-			}),
-		);
-
-		await waitFor(() =>
-			expect(screen.getByText('The admin wallet address field is required.')).not.toBe(null),
+			expect(
+				screen.getByText('The symbol field is required.'),
+			).toBeInTheDocument(),
 		);
 	});
 
@@ -105,7 +88,7 @@ describe('<TokenomicsForm />', () => {
 		render(
 			<ZUIProvider>
 				<CreateTokenFormContext.Provider value={DEFAULT_PROVIDER_VALUES}>
-					<TokenomicsForm {...DEFAULT_PROPS} />
+					<DetailsForm {...DEFAULT_PROPS} />
 				</CreateTokenFormContext.Provider>
 			</ZUIProvider>,
 		);
@@ -120,32 +103,24 @@ describe('<TokenomicsForm />', () => {
 		render(
 			<ZUIProvider>
 				<CreateTokenFormContext.Provider value={DEFAULT_PROVIDER_VALUES}>
-					<TokenomicsForm {...DEFAULT_PROPS} />
+					<DetailsForm {...DEFAULT_PROPS} />
 				</CreateTokenFormContext.Provider>
 			</ZUIProvider>,
 		);
 
 		const user = userEvent.setup();
 
-		await user.type(screen.getByPlaceholderText(/Enter total supply.../i), '123');
-		
-		await user.type(
-			screen.getByPlaceholderText(/Enter initial wallet address.../i),
-			'0xb794f5ea0ba39494ce839613fffba74279579268',
-		);
-
-		await user.type(
-			screen.getByPlaceholderText(/Enter admin wallet address.../i),
-			'0xb794f5ea0ba39494ce839613fffba74279579268',
-		);
+		await user.type(screen.getByPlaceholderText(/Enter name.../i), 'Test');
+		await user.type(screen.getByPlaceholderText(/Enter symbol.../i), 'TEST');
 
 		await user.click(screen.getByRole('button', { name: /Next/i }));
 
 		await waitFor(() =>
 			expect(onSubmit).toHaveBeenCalledWith({
-				tokenCount: '123',
-				initialTokenSupplyWalletAddress: '0xb794f5ea0ba39494ce839613fffba74279579268',
-				adminWalletAddress: '0xb794f5ea0ba39494ce839613fffba74279579268'
+				mediaType: undefined,
+				previewUrl: '',
+				name: 'Test',
+				symbol: 'TEST',
 			}),
 		);
 	});
