+++ src/features/create-token/Steps/TokenomicsForm.test.tsx
@@ -4,17 +4,17 @@ import userEvent from '@testing-library/user-event';
 import { fireEvent, render, waitFor, screen } from '@testing-library/react';
 
 import { ZUIProvider } from '@zero-tech/zui/ZUIProvider';
-import { DetailsForm, DetailsFormProps } from './';
+import { TokenomicsForm, TokenomicsFormProps } from './';
 import { CreateTokenFormContext } from '../';
 
 let onSubmit = jest.fn();
 
-const DEFAULT_PROPS: DetailsFormProps = {
+const DEFAULT_PROPS: TokenomicsFormProps = {
 	onClose: jest.fn(),
 };
 
 const DEFAULT_PROVIDER_VALUES = {
-	stepId: 'details',
+	stepId: 'tokenomics',
 	title: 'Create Token',
 	details: {
 		mediaType: undefined,
@@ -30,24 +30,24 @@ const DEFAULT_PROVIDER_VALUES = {
 	onStepUpdate: jest.fn(),
 	onTitleUpdate: jest.fn(),
 	onDetailsChange: jest.fn(),
-	onDetailsSubmit: onSubmit,
-	onTokenomicsSubmit: jest.fn(),
+	onDetailsSubmit: jest.fn(),
+	onTokenomicsSubmit: onSubmit,
 	onLaunchSubmit: jest.fn(),
 };
 
-describe('<DetailsForm />', () => {
+describe('<TokenomicsForm />', () => {
 	beforeEach(() => jest.resetAllMocks());
 
-	test('should correctly validate required name field', async () => {
-		render (
+	test('should correctly validate required token count field', async () => {
+		render(
 			<ZUIProvider>
 				<CreateTokenFormContext.Provider value={DEFAULT_PROVIDER_VALUES}>
-					<DetailsForm {...DEFAULT_PROPS} />
+					<TokenomicsForm {...DEFAULT_PROPS} />
 				</CreateTokenFormContext.Provider>
-			</ZUIProvider>
+			</ZUIProvider>,
 		);
 
-		fireEvent.blur(screen.getByPlaceholderText(/Enter name.../i));
+		fireEvent.blur(screen.getByPlaceholderText(/Enter total supply.../i));
 		fireEvent.click(
 			screen.getByRole('button', {
 				name: 'Next',
@@ -55,20 +55,24 @@ describe('<DetailsForm />', () => {
 		);
 
 		await waitFor(() =>
-			expect(screen.getByText('The name field is required.')).not.toBe(null),
+			expect(screen.getByText('The token count field is required.')).not.toBe(
+				null,
+			),
 		);
 	});
 
-	test('should correctly validate required symbol field', async () => {
-		render (
+	test('should correctly validate required initial wallet address field', async () => {
+		render(
 			<ZUIProvider>
 				<CreateTokenFormContext.Provider value={DEFAULT_PROVIDER_VALUES}>
-					<DetailsForm {...DEFAULT_PROPS} />
+					<TokenomicsForm {...DEFAULT_PROPS} />
 				</CreateTokenFormContext.Provider>
-			</ZUIProvider>
+			</ZUIProvider>,
 		);
 
-		fireEvent.blur(screen.getByPlaceholderText(/Enter symbol.../i));
+		fireEvent.blur(
+			screen.getByPlaceholderText(/Enter initial wallet address.../i),
+		);
 		fireEvent.click(
 			screen.getByRole('button', {
 				name: 'Next',
@@ -76,7 +80,34 @@ describe('<DetailsForm />', () => {
 		);
 
 		await waitFor(() =>
-			expect(screen.getByText('The symbol field is required.')).not.toBe(null),
+			expect(
+				screen.getByText('The initial wallet address field is required.'),
+			).toBeInTheDocument(),
+		);
+	});
+
+	test('should correctly validate required admin wallet address field', async () => {
+		render(
+			<ZUIProvider>
+				<CreateTokenFormContext.Provider value={DEFAULT_PROVIDER_VALUES}>
+					<TokenomicsForm {...DEFAULT_PROPS} />
+				</CreateTokenFormContext.Provider>
+			</ZUIProvider>,
+		);
+
+		fireEvent.blur(
+			screen.getByPlaceholderText(/Enter admin wallet address.../i),
+		);
+		fireEvent.click(
+			screen.getByRole('button', {
+				name: 'Next',
+			}),
+		);
+
+		await waitFor(() =>
+			expect(
+				screen.getByText('The admin wallet address field is required.'),
+			).toBeInTheDocument(),
 		);
 	});
 
@@ -84,11 +115,11 @@ describe('<DetailsForm />', () => {
 		render(
 			<ZUIProvider>
 				<CreateTokenFormContext.Provider value={DEFAULT_PROVIDER_VALUES}>
-					<DetailsForm {...DEFAULT_PROPS} />
+					<TokenomicsForm {...DEFAULT_PROPS} />
 				</CreateTokenFormContext.Provider>
 			</ZUIProvider>,
 		);
-		
+
 		const user = userEvent.setup();
 		await user.click(screen.getByRole('button', { name: /Next/i }));
 
@@ -99,27 +130,36 @@ describe('<DetailsForm />', () => {
 		render(
 			<ZUIProvider>
 				<CreateTokenFormContext.Provider value={DEFAULT_PROVIDER_VALUES}>
-					<DetailsForm {...DEFAULT_PROPS} />
+					<TokenomicsForm {...DEFAULT_PROPS} />
 				</CreateTokenFormContext.Provider>
 			</ZUIProvider>,
 		);
 
 		const user = userEvent.setup();
 
-		await user.type(screen.getByPlaceholderText(/Enter name.../i), 'Test');
 		await user.type(
-			screen.getByPlaceholderText(/Enter symbol.../i),
-			'TEST',
+			screen.getByPlaceholderText(/Enter total supply.../i),
+			'123',
+		);
+
+		await user.type(
+			screen.getByPlaceholderText(/Enter initial wallet address.../i),
+			'0x29D7d1dd5B6f9C864d9db560D72a247c178aE86B',
+		);
+
+		await user.type(
+			screen.getByPlaceholderText(/Enter admin wallet address.../i),
+			'0x29D7d1dd5B6f9C864d9db560D72a247c178aE86B',
 		);
 
 		await user.click(screen.getByRole('button', { name: /Next/i }));
 
 		await waitFor(() =>
 			expect(onSubmit).toHaveBeenCalledWith({
-				mediaType: undefined,
-				previewUrl: '',
-				name: 'Test',
-				symbol: 'TEST',
+				tokenCount: '123',
+				initialTokenSupplyWalletAddress:
+					'0x29D7d1dd5B6f9C864d9db560D72a247c178aE86B',
+				adminWalletAddress: '0x29D7d1dd5B6f9C864d9db560D72a247c178aE86B',
 			}),
 		);
 	});
